+++ Source/AccelByteNetworkUtilities/Private/Networking/AccelByteJuice.h
@@ -11,6 +11,7 @@
 #include "Dom/JsonObject.h"
 #include "Containers/Queue.h"
 #include "Models/AccelByteTurnModels.h"
+#include "Core/AccelByteDefines.h"
 
 /*
  * Currently max char host name, username, and password is 64.
@@ -73,6 +74,22 @@ public:
 	*/
 	virtual bool IsPeerReady() const override;
 
+	/**
+	 * Schedule reconnection to host in the future, this is to make sure libjuice agent in host already closed before
+	 * initiating new connection.
+	 *
+	 * @param TurnServerCredential Turn server credential tobe used in reconnection
+	 * @param InNextSeconds the reconnection will start after this seconds from the time peer inactive timeout
+	 */
+	virtual void ScheduleReconnection(const FAccelByteModelsTurnServerCredential& TurnServerCredential, float InNextSeconds = 1) override;
+
+	
+	/**
+	 * [FOR GAUNTLET TEST ONLY] simulate network switching by ignoring update to LastPeerActivityTime.
+	 * Later it will trigger connection lost and reconnection.
+	 */
+	virtual void SimulateNetworkSwitching() override;
+
 private:
 	// instance of the Juice agent, will handle the ICE connection
 	juice_agent_t *JuiceAgent = nullptr;
@@ -126,6 +143,27 @@ private:
 	 */
 	juice_state_t LastJuiceState = JUICE_STATE_DISCONNECTED;
 
+	/*
+	 * Flag indicating reconnection for replacing libjuice agent with a new one.
+	 * When this flag set to true the on connected delegate shouldn't be triggered to avoid messing up OSS.
+	*/
+	bool bReconnection = false;
+
+	// This time indicate the last time this agent received a data from its peer
+	TAtomic<double> LastPeerActivityTime{0};
+
+	// Interval for checking peer last activity ticker in seconds
+	float PeerLastActivityTickerIntervalInSeconds = 1.0f;
+
+	// Maximum peer inactive time, counted from the last time the agent received data.
+	double PeerInactiveTimeoutInSeconds = 10.0;
+
+	// Peer last activity delegate and delegate handle
+	FTickerDelegate PeerLastActivityTickerDelegate;
+	FDelegateHandleAlias PeerLastActivityTickerDelegateHandle;
+
+	bool bSimulateNetworkSwitchingEnabled = false; 
+
 	/**
 	 * @brief Create peer connection using defined Turn server
 	 *
@@ -168,9 +206,38 @@ private:
 	/**
 	 * @brief Update peer status based on message received from signaling
 	 *
-	 * @param Message message string from signaling
+	 * @param InStatus Peer status, true if hosting.
 	 */
 	void UpdatePeerStatus(bool InStatus);
+
+	/**
+	 * Update LastPeerActivityTime to current time
+	 */
+	void UpdatePeerLastActivity();
+
+	/**
+	 * Method attached to core ticker for initiating reconnection to host.
+	 *
+	 * @param DeltaTime elapsed delta time in seconds.
+	 * @return return false because only used as one shot timer.
+	 */
+	bool ReconnectOnNextTick(float DeltaTime);
+
+	/**
+	 * Method attached to core ticker for monitoring peer last activity time. This method start to run regularly from
+	 * the time its connected to peer. It will stop when peer inactive timeout detected and schedule next attempt to
+	 * reconnect to host. In host side it only make sure the libjuice agent closed and destroyed.
+	 *
+	 * @param DeltaTime elapsed delta time in seconds.
+	 * @return return true to restart the timer and false to stop it.
+	 */
+	bool PeerLastActivityTick(float DeltaTime);
+
+	/**
+	 * Start a ticker to monitor peer last activity time, called after peer connected.
+	 * The interval is set by PeerLastActivityTickerIntervalInSeconds
+	 */
+	void StartPeerLastActivityWatcher();
 };
 
 #endif